Test suite is not working with typescript
