+++ decidim-comments/app/frontend/comments/comment_order_selector.component.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import { I18n }   from 'react-i18nify';
+
+const { I18n } = require('react-i18nify');
 
 interface CommentOrderSelectorProps {
   defaultOrderBy: string;
