+++ karma.conf.js
@@ -2,20 +2,20 @@ process.env.BABEL_ENV = 'test';
 const webpackEnv = { test: true };
 const webpackConfig = require('./webpack.config')(webpackEnv);
 
-const testGlob = 'decidim-*/app/frontend/entry.test.js';
+const testGlob = 'decidim-*/app/frontend/entry.test.ts';
 const srcGlob = 'decidim-*/app/frontend/**/!(*.test.*)';
 
 module.exports = function(config) {
   config.set({
     basePath: '',
     frameworks: ['jasmine'],
-    files: [testGlob, srcGlob, 'decidim-core/**/*.component.js.es6', 'decidim-core/**/*.test.js'],
-    exclude: ['decidim-*/app/frontend/entry.js'],
+    files: [testGlob, srcGlob], //'decidim-core/**/*.component.js.es6', 'decidim-core/**/*.test.js'],
+    exclude: ['decidim-*/app/frontend/entry.ts'],
     preprocessors: {
       [testGlob]: ['webpack', 'sourcemap'],
       [srcGlob]: ['webpack'],
-      'decidim-core/**/*.test.js': ['webpack', 'sourcemap'],
-      'decidim-core/**/*.component.js.es6': ['webpack']
+      // 'decidim-core/**/*.test.js': ['webpack', 'sourcemap'],
+      // 'decidim-core/**/*.component.js.es6': ['webpack']
     },
     webpack: webpackConfig,
     webpackMiddleware: { noInfo: true },
