+++ package.json
@@ -20,6 +20,18 @@
     "postgraphql:generate_schema_types": "rm schema.json"
   },
   "devDependencies": {
+    "@types/chai": "^3.4.35",
+    "@types/chai-as-promised": "^0.0.29",
+    "@types/chai-enzyme": "^0.5.7",
+    "@types/classnames": "^0.0.32",
+    "@types/core-js": "^0.9.35",
+    "@types/enzyme": "^2.7.5",
+    "@types/faker": "^3.1.32",
+    "@types/jasmine": "^2.5.43",
+    "@types/react": "^15.0.12",
+    "@types/react-dom": "^0.14.23",
+    "@types/sinon-chai": "^2.7.27",
+    "@types/uuid": "^2.0.29",
     "awesome-typescript-loader": "^3.0.4",
     "babel-core": "~6.23.1",
     "babel-eslint": "~7.1.1",
@@ -33,7 +45,8 @@
     "chai": "~3.5.0",
     "chai-as-promised": "~6.0.0",
     "chai-enzyme": "~0.6.1",
-    "enzyme": "~2.7.1",
+    "chai-jasmine": "^1.0.7",
+    "enzyme": "^2.7.1",
     "eslint": "~3.15.0",
     "eslint-loader": "~1.6.1",
     "eslint-plugin-import": "~2.2.0",
@@ -41,10 +54,12 @@
     "eslint-plugin-react": "~6.10.0",
     "expose-loader": "~0.7.1",
     "faker": "~3.1.0",
+    "jasmine": "^2.5.3",
     "jasmine-core": "~2.5.2",
+    "jasmine-enzyme": "^2.1.2",
     "json-loader": "~0.5.4",
     "karma": "~1.4.1",
-    "karma-chai": "~0.1.0",
+    "karma-chai": "^0.1.0",
     "karma-coverage": "~1.1.1",
     "karma-jasmine": "~1.1.0",
     "karma-phantomjs-launcher": "~1.0.2",
@@ -54,10 +69,9 @@
     "progress-bar-webpack-plugin": "~1.9.0",
     "raw-loader": "~0.5.1",
     "react-addons-test-utils": "~15.4.0",
-    "sinon": "~1.17.6",
     "sinon-chai": "~2.8.0",
     "source-map-loader": "^0.1.6",
-    "typescript": "^2.1.6",
+    "typescript": "2.1.x",
     "webpack": "~2.2.1",
     "webpack-bundle-analyzer": "~2.3.0",
     "webpack-config-utils": "~2.3.0",
@@ -81,6 +95,7 @@
     "react-apollo": "^1.0.0-rc.2",
     "react-dom": "~15.4.2",
     "react-i18nify": "~1.8.4",
+    "sinon": "^1.17.7",
     "uuid": "~3.0.1"
   }
 }
