+++ webpack.d.ts
@@ -1,5 +1,4 @@
 declare var $: any;
-declare var require: any;
 
 declare interface Window {
     Comments: any;
